pyproject.toml
[project]
name = "whisper-transcription"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.9.10"
dependencies = [
    "aiofiles>=24.1.0",
    "ffmpeg>=1.4",
    "openai>=1.59.9",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
]
